[Core,Jobs,Geo] refactored Jobs\Entity\Location, using Core\Entity\AbstractLocation class as base class for Jobs\Entity\Location and Cv\Entity\Location
