+++ module/Jobs/src/Jobs/Entity/Location.php
@@ -10,7 +10,7 @@
 /** Location.php */
 namespace Jobs\Entity;
 
-use Core\Entity\AbstractEntity;
+use Core\Entity\AbstractLocation;
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
 use GeoJson\GeoJson;
 
@@ -21,143 +21,6 @@ use GeoJson\GeoJson;
  * @ODM\Index(keys={"coordinates"="2dsphere"})
  *
  */
-class Location extends AbstractEntity implements LocationInterface
+class Location extends AbstractLocation
 {
-    /**
-     * city name of a job location
-     *
-     * @ODM\Field(type="string")
-     */
-    protected $city;
-
-    /**
-     * region of a job location. E.g "Hessen" is a region in germany
-     *
-     * @ODM\Field(type="string")
-     */
-    protected $region;
-
-    /**
-     * postalcode of a job location.
-     *
-     * @var String
-     * @ODM\Field(type="string")
-     */
-    protected $postalcode;
-
-    /**
-     * coordinates of a job location.
-     *
-     * @var GeoJson
-     * @ODM\EmbedOne(discriminatorField="_entity")
-     */
-    protected $coordinates;
-
-    /**
-     * Country of a job location
-     * @var String
-     *
-     * @ODM\Field(type="string")
-     */
-    protected $country;
-    
-    public function __construct()
-    {
-    }
-
-    public function preUpdate()
-    {
-    }
-
-    public function getCoordinates()
-    {
-        return $this->coordinates;
-    }
-
-    /**
-     * @param GeoJson $coordinates
-     *
-     * @return $this
-     */
-    public function setCoordinates(GeoJson $coordinates)
-    {
-        $this->coordinates = $coordinates;
-        return $this;
-    }
-
-    /**
-     * @return String
-     */
-    public function getPostalcode()
-    {
-        return $this->postalcode;
-    }
-
-    /**
-     * @param $postalcode
-     *
-     * @return $this
-     */
-    public function setPostalcode($postalcode)
-    {
-        $this->postalcode = $postalcode;
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getCity()
-    {
-        return $this->city;
-    }
-
-    /**
-     * @param $country
-     *
-     * @return $this
-     */
-    public function setCity($country)
-    {
-        $this->city = $country;
-        return $this;
-    }
-
-    /**
-     * @return String
-     */
-    public function getCountry()
-    {
-        return $this->country;
-    }
-
-    /**
-     * @param $country
-     *
-     * @return $this
-     */
-    public function setCountry($country)
-    {
-        $this->country = $country;
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getRegion()
-    {
-        return $this->region;
-    }
-
-    /**
-     * @param $region
-     *
-     * @return $this
-     */
-    public function setRegion($region)
-    {
-        $this->region = $region;
-        return $this;
-    }
 }
